+++ sdk/metricsadvisor/Azure.AI.MetricsAdvisor/src/Models/NotificationHook/NotificationHook.cs
@@ -3,6 +3,7 @@
 
 using System;
 using System.Collections.Generic;
+using System.Text.Json;
 using Azure.AI.MetricsAdvisor.Models;
 using Azure.Core;
 
@@ -12,7 +13,7 @@ namespace Azure.AI.MetricsAdvisor.Administration
     /// An alert notification to be triggered after an anomaly is detected by Metrics Advisor.
     /// </summary>
     [CodeGenModel("HookInfo")]
-    public partial class NotificationHook
+    public abstract partial class NotificationHook
     {
         internal NotificationHook(string name)
         {
@@ -85,15 +86,88 @@ namespace Azure.AI.MetricsAdvisor.Administration
                         Headers = h.Headers
                     }
                 },
-                _ => throw new InvalidOperationException("Unknown hook type.")
+                _ => new HookInfoPatch()
             };
 
+            patch.HookType = hook.HookType;
             patch.HookName = hook.Name;
             patch.Description = hook.Description;
             patch.ExternalLink = hook.ExternalUri?.AbsoluteUri;
             patch.Admins = hook.AdministratorEmails;
 
             return patch;
+        }
+
+        internal static NotificationHook DeserializeNotificationHook(JsonElement element)
+        {
+            if (element.TryGetProperty("hookType", out JsonElement discriminator))
+            {
+                switch (discriminator.GetString())
+                {
+                    case "Email":
+                        return EmailNotificationHook.DeserializeEmailNotificationHook(element);
+                    case "Webhook":
+                        return WebNotificationHook.DeserializeWebNotificationHook(element);
+                }
+            }
+            HookType hookType = default;
+            Optional<string> hookId = default;
+            string hookName = default;
+            Optional<string> description = default;
+            Optional<string> externalLink = default;
+            Optional<IReadOnlyList<string>> admins = default;
+            foreach (var property in element.EnumerateObject())
+            {
+                if (property.NameEquals("hookType"))
+                {
+                    hookType = new HookType(property.Value.GetString());
+                    continue;
+                }
+                if (property.NameEquals("hookId"))
+                {
+                    hookId = property.Value.GetString();
+                    continue;
+                }
+                if (property.NameEquals("hookName"))
+                {
+                    hookName = property.Value.GetString();
+                    continue;
+                }
+                if (property.NameEquals("description"))
+                {
+                    description = property.Value.GetString();
+                    continue;
+                }
+                if (property.NameEquals("externalLink"))
+                {
+                    externalLink = property.Value.GetString();
+                    continue;
+                }
+                if (property.NameEquals("admins"))
+                {
+                    if (property.Value.ValueKind == JsonValueKind.Null)
+                    {
+                        property.ThrowNonNullablePropertyIsNull();
+                        continue;
+                    }
+                    List<string> array = new List<string>();
+                    foreach (var item in property.Value.EnumerateArray())
+                    {
+                        array.Add(item.GetString());
+                    }
+                    admins = array;
+                    continue;
+                }
+            }
+            return new UnknownNotificationHook(hookType, hookId.Value, hookName, description.Value, externalLink.Value, Optional.ToList(admins));
+        }
+
+        private class UnknownNotificationHook : NotificationHook
+        {
+            public UnknownNotificationHook(HookType hookType, string id, string name, string description, string internalExternalLink, IReadOnlyList<string> administrators)
+                : base(hookType, id, name, description, internalExternalLink, administrators)
+            {
+            }
         }
     }
 }