+++ sdk/metricsadvisor/Azure.AI.MetricsAdvisor/tests/Models/DataSourceCredentialEntitiesTests.cs
@@ -2,10 +2,7 @@
 // Licensed under the MIT License.
 
 using System;
-using System.IO;
 using System.Linq;
-using System.Text;
-using System.Threading;
 using System.Threading.Tasks;
 using Azure.AI.MetricsAdvisor.Administration;
 using Azure.AI.MetricsAdvisor.Models;
@@ -83,7 +80,7 @@ namespace Azure.AI.MetricsAdvisor.Tests
             MockRequest request = mockTransport.Requests.First();
             string content = ReadContent(request);
 
-            Assert.That(content, Contains.Substring("\"accountKey\":\"secret\""));
+            Assert.That(content, ContainsJsonString("accountKey", "secret"));
         }
 
         [Test]
@@ -107,7 +104,7 @@ namespace Azure.AI.MetricsAdvisor.Tests
             MockRequest request = mockTransport.Requests.First();
             string content = ReadContent(request);
 
-            Assert.That(content, Contains.Substring("\"clientSecret\":\"secret\""));
+            Assert.That(content, ContainsJsonString("clientSecret", "secret"));
         }
 
         [Test]
@@ -131,7 +128,7 @@ namespace Azure.AI.MetricsAdvisor.Tests
             MockRequest request = mockTransport.Requests.First();
             string content = ReadContent(request);
 
-            Assert.That(content, Contains.Substring("\"keyVaultClientSecret\":\"secret\""));
+            Assert.That(content, ContainsJsonString("keyVaultClientSecret", "secret"));
         }
 
         [Test]
@@ -155,15 +152,7 @@ namespace Azure.AI.MetricsAdvisor.Tests
             MockRequest request = mockTransport.Requests.First();
             string content = ReadContent(request);
 
-            Assert.That(content, Contains.Substring("\"connectionString\":\"secret\""));
-        }
-
-        private string ReadContent(Request request)
-        {
-            using MemoryStream stream = new MemoryStream();
-            request.Content.WriteTo(stream, CancellationToken.None);
-
-            return Encoding.UTF8.GetString(stream.ToArray());
+            Assert.That(content, ContainsJsonString("connectionString", "secret"));
         }
     }
 }